fix: rl

+++ backend/internal/rateLimiter/ratelimiter.go
@@ -8,12 +8,12 @@ import (
 type RateLimiter struct {
 	visitors map[string]*Visitor
 	mu       sync.Mutex
-	rate     int
+	rate     float64
 	burst    int
 	period   time.Duration
 }
 
 type Visitor struct {
-	tokens      int
+	tokens      float64
 	lastRequest time.Time
 }
